parking-management-system/requirements.txt
opencv-python
ultralytics
pytesseract
pyserial
